feat: move relay data types and base state to commons (batch 9)
Move BasicRelaySetupInfo, RelayCountResult, CountFilter data classes
and replace() extension to commons for KMP.

Create BasicRelaySetupInfoState abstract ViewModel in commons that
captures the platform-agnostic relay list state management
(add/delete/move/clear, count loading). Android BasicRelaySetupInfoModel
extends it, providing Amethyst-specific implementations for relay info
loading, tor evaluation, and signer access.

Part of the KMP iOS migration tracked in #2238.
